Added a basic Docker setup
To have a working lokal database that is like the production effortelss,
I added this small docker setup. It just hosts a MariaDB instance that
can be accessed through the localhost.

Added the `.example.env` for showcasing how the required `.env` file
needs to look like.

Reworked the readme, to be more structured and not a complete
cluster-fuck with my added Docker setup.