Rudimentary Docker setup for Python
The database is already running in Docker and does so also for
development.

The now added python service, called `p` as the project, hoewever is
only meant for deployment and does not feature any commodities for
develpment. This is mostly due to Python being extremly frustrating to
dockerize. A simple volume mount does not do it, at least not to my
limited python knowledge.