Volumes for the Docker setup
Added a few volumes, so the code is directly mounted into the Docker
container. This removes the necessity to rebuild the image on code
changes. It might not be the most performant way, but we are more for
convenience at the moment.

I updated the readme with details on how to work with the setup now.