More recent npm version for Docker image
Using Debian we do not get the latest nodejs and npm version using their
built in package manager. This usually works, but on Jan's laptop it now
gets stuck installing npm packages.

So I think installing a real version of nodejs and npm might be a
solution.

This is what I added here. Followed the instructions from the official
website.