Improved logging
The logging is still configured in the `__init__.py` but we no longer
also load the loggers there.

The loggers are loaded in the files where they are needed. I think this
is the way they are supposed to be used. At least this is the way I am
seeing in every tutorial.