Installed a mail server for testing
Here, I added a new docker compose service, called mailhog. This service
is a dummy mail server that our app can use to send mails and we to test
how they look and if they go out at all.

This will help me with setting up and debug the mailing in the code.