Fixed `env.py` missing
I didn't realize the generated `env.py` was not checked in when I
installed the migrations library.

It got excluded due to a rule in the `.gitignore`. I kept the rule but
created an exception for that one `env.py` file.

I also thoroughly checked the `env.py` but could not fing any sensible
information. So it seems safe to be checked in.