Fixed GET routes
In the MR !45 I noticed how we had two functions for the same route. One
for GET and the other for POST requests presumably. However, the
function for the POST requests had the GET method registered on it as
well, making the other function superfluous.

Here, I fixed this by giving back the GET requests to the functions that
should handle them and leaving the POST requests to the ones designed to
do so.

I decided against removing the superfluous functions all together, as I
think it makes the POST functions a bit easier to read. One function
should fulfill exactly one job.