Fixed reading boolean environment variables
Debugging the mailing, I found that we do not parste the environment
variables correctly. They come in as strings and thus are not
interpreted as the boolean values we set.

The solution I implemented here explicitly compares a given environment
variable against a set of truthy strings to determine its boolean value.