Renamed env vars to new app name 'Voy'
Digging into the dokcer setup when working on the mailing, I found that
we are still using the old name for the app in the environment variables
and in the docker setup.

Here I adapted all the old names.

I kept the `project-p` namespace for the docker-images. Maybe this
should be renamed to KAIZN.