Added Make target for initializing the database
I added a Make target for executing the `voy database init` command and
referenced it in the readme.

This is more of documentation for how to work with our project during
development or deployment, than actually improving workflows. For
experienced users the make targets might not serve any purpose, but for
users just getting started it can be used as a reference on what is
possible and how to do it.