Updated the readme
With the new Make targets, the readme needs to be adapted. Doing so I
realized, that it doesn't just need adaptation, but an overhaul.

So here I overhauled the development part of the readme. It is shorter
now and together with the quickstart a good introduction to developing
voy.

With that done, I had to update the Deployment section as well, as it
dependet heavily on the Development section.

I also renamed headline here and there as I saw fit.