Updated parcel
This not only fixes our bundling, but also resolves the issue #33
concerning insecure npm packages.

I used `npm audit fix --force` to fix all issues with our packages. It
installed the beta of the new parcel version 2. I had to adapt the npm
scripts as well as add a `.parcelrc` file to make things work again. For
faster builds and development, I removed the `--no-cache` flag and thus
had to ignore the cache folder created by parcel.
I also threw out the `--no-minify` flag. I am unsure why we decided to
implement it. I think it was out of simplicity or for debugging the
production setup. We don't need that anymore.