Removed docker setup env override
Beofre, the docker setup overrid some of the environment variables. This
was for the ease of use of the docker setup. But now, we are more
experienced in working with docker and need more flexibility. Also, I
like configuration over convention in this case. It is more transparent
I feel.

I dapted the `.example.env` file, so that it works right out of the box.