Chain commands with double ampersand
Here, I updated the make target for setting up the virtual environment.
Before, it was using semicolons to chain the commands for that. This has
the downside of executing the next command regardless of the previous
command failig.

I replaced the semicolons with double ampersands (&&) so that the are
only executed if the previous command was successfull. This is important
as they are all dependet on another.