Dependencies in `setup.py`
In this commit, I moved all direct dependencies to the `setup.py`. Doing
so, I also removed the pinned versions, and allowed for minor version
upgrades. Then I reinstalled everything and updated the
`requirements.txt`.

This makes it easier to manage the packages required for voy. Because
now we know which packages are direct requirements and which are
indirect ones. No only makes it easier to add and remove packages, but
also to update them.