Fixed package installation
Before, not all files were included when installing the voy package.
This lead to the package not functioning when installed via `pip install .`.

I now added configuration to make sure all the necessary files are
copied or included in the package when installing it.

For that I had to add another `__init__py` into the `data_analysis`
folder. So that it would be recognized by pip as a module.