Replaced Parcel with Symmfony Webpack Encore
With Parcel not working on arm64 architecture, we needed another
bundler. I tried Rollup - which worked, but configuring it was tiresome.
So I ditched it and went with Symfony Webpack Encore. A very well built
bundler configuration that relies on the most common bundler, Webpack,
which I already know from other projects. Easy choice. Why didn't I take
that right away?

Works on my emulated arm64 architecture. So it should work in Docker on
Apples M1 as well.

I basically did a default install of Encore, and adapted it so that we
get everything in a single file without hashes. Then I adapted some
paths throughout the project.

Everything seems to work now.