More readable error and success messages for pw reset command
I noticed that long sentences on the console are not that easy to read.
Better said require effort to read. A simple word like "Error" or
"Success" is much much faster to grasp.

So I rewrote the error and success messages of the password reset
command, so that they always start with an simple and easy word.
After that, there are morde detailes included.

Also I found adding new lines before and after the output helpes with
readability even more as the output is now clearly separated from
everything else on the command line.