Moved admin user creation to database initialization
I think it makes more sense to create the admin user at the same time,
when the database is initialized. This way we can make sure that no one
can intercept the admin creation, and hijack the system.

Also it makes for a nicer setup process. The console asks you for the
data required. And you no longer get that whacky message on initial
login.