Moved QueriesTags model into separate file
My aim is to have one model per file. I think this makes it easier to
find models.

So here, I separated the QueriesTags model from the Queries model to
achieve just that.

I also renamed the QueriesTags model from QueryTags to indicate that
these are tags specifically for the Queries.