Fixed Studies to DB_User relationship
Turns out we need both for it to work:
- An import for the Studies model in the `__init__.py`.
- `back_populates` rather than `backref`.

I think the first is because otherwise python would not load the model
and the code in the studies.py is never even executed. Makes sense that
SQLAlchemy can't find it then.

The latter is apparently because `backref` only works when you don't
have an extra relationship table defined like we do.