Extracted study controller
I extracted the actions (functions in the controller) for adding a
new study from the qc_database controller.

This controller is bloated and needs to be split up.

I also added a folder in the templates for that new study controller.

Furthermore, I adapted the routes to be namespaced. This will help
building a clear url structure and add futher pages in the future.