Fixed flashes (on login page)
On the login page, we no longer had flashes after the merge with
make-voy-beatiful-again.

I fixed this by adding flashes to the login template. Flask seems to
track when they are shown. So showing them on the login page resolves
this issue. Also there are flashes that need to be shown there, so that
is fixed as well.