Extracted footer content into own template
This is technically not necessary. But I think it makes the Code more
readable, less cluttered and more reusable.

I try to think with templates just like I think with functions: Every
one should just serve one specific purpose. So clearly the layout does
the layouting (the arrangement of all the parts of the page), while the
footer content just does that, it renders the footer content.

Nice separation of concerns.