Very basic page layout with navigation on the left
Built a very basic page layout. This layout should be used for most
pages in voy some day.

I put the navigation to the left as I think this gives us more space for
the menu and other buttons / items.

So far, the menu only contains dummy content as I just wanted to get the
layout right. Not the functionality.