Added block for modals
I added a block for modals and other "on-demand" page content that is
typically not visible right away.
Modals will be appended to the very end of the page, right above the
footer scripts.

I added this block so one can easily add modals to pages, without
worrying about their placement. in the markup. And this might make
styling easier as one does not need to add additional wrappers jus to
separate the page content from the modal content.