Created a macro for rendering a section
I want to reuse this section as much as possible, without copying the
code. This would make adaptions harder and annoying.

So I created this macro to render a section. Was a bit tricky as Jinja
does not seem to support anything like embeds from Twig or slots from
Vue. So I had to come up with my own strategy of how to pass multiple
pieces of markup into that macro.
I think I found an ok way with using the caller strategy for the main
content piece and a set-block for everything else. Together with a
with-block this should work just fine and not pollute my global
namespace.

Not too happy, but I can work with that.