Optional section title
I made the section title optional, so that you can create a section
without title.

I want to use this for pages where there is only one section. For such
pages, the page title will already serve as a title for that section, so
another section-title is not needed in such cases.