Basic Alpine modal
Started implementing the ticket comment modal. Added all the basic files
I need for that. As well as some dummy code. Just to have something pop
up and see if everything works.

So far the modal opens and closes with a nice opacity transition. Was a
bit tricky to get this to work. Bulma and Alpine have different ways of
showing modals, so I had to come up with a way that works. It now works
by having the modal always active, that is visible in terms of Bulmas
styling and then show and hide it via Alpine.

Furthermore, I had to shuffle around the imports in the `main.js`.
Otherwise my Alpine component for the modal would not work. Also, I
didn't want to manually import each and every component into the
`alpinejs.js` file. I wanted adding Alpine components to be easy and not
requiring the knowledge where exactly to import them to make them work.