Wait for the ticket comment modal to close before resetting
I noticed, that the modal was resetting before it was no longer visible.
This way, the user could see the ticket id switching to `undefined`
before the modal was gone.
I don't like such things. It makes the software feel cheap and
errornous.

So here, I added logic to wait for the css transitions to complete
before resetting the content of the modal.