Added request for loading the comments
Slowly the ticket comment modal gets to where it should be. I now added
a request to load the markup for the rendered comments from the backend.

Added two new methods for that. One that soley does the loading, one
that triggers the loading and then updates the content.

Also, I added some styling to give the modal a fixed height. This
prevents any jumping, when the comment markup is loaded an inserted.