More precise show and hide promises
Here, I improved the show and hide promises of the ticket comment modal.
They now wait for alpine to render the changes of the modal state and
then retur the promise that waits for the css transition.

For that I:
- Added a new waitForRender method that wraps the $nextTick function from alpine with
  a promise.
- Made use of that new waitForRender method to time the opening and
  closing more precise.

This is not really necessary for shoing and hiding the modal. But I
really need this waitForRender method for loading and displaying the
comments in a user-friendly way. That will come in the next commits.