Jump to the most recent comment initially
To not make the user scroll all the way down to the most recent comment,
the ticket comment modal now scrolls to the most recent comment
automatically.

For that, I had to rework the way, the modal is shown. I can't set the
scrollTop on elements that are not displayed. I can, but it has no
effect.
So now, the modal is displayed right away, but 100% transparent. This
way I can initialize the invisible modal, and then fade it in
afterwards.

For that I:
- Added a few helper classes that transition the modal (or anything else).
- Differentiated between open and visible in the modal code.
- Added a ref to the element containing the comments.
- Added a method that uses that ref to scroll it all the way to the
  bottom.