Debounced comment loading
To not get into some race conditions, I debounced the loading of the
comments in the ticket comment modal.

Now, still active requests are canceled, when the user closes the modal.
This way the modals state is not updated accidentally after it closed.
Same with loading comments. Ongoing requests are canceled before making
a new one. This is just a good practice.