Form to add comment in the ticket commemt modal
Finally, I could add the last piece of functionality: Adding a comment.

I added markup for the form. It has a Alpine reference, so that I can
use it in the JavaScript code and the submit event is handled in
JavaScript as well.

Added a method for making the call to the backend to actually save the
new comment.

Added another method that cares about locking the ui, updating the
comments and scrolling to the most recent one.

Also, I extracted the code for loading and displaying comments into a
separate function. This way I can reuse that function when initializing
the comments or when adding a new one and reloading them.