Reworked the users index to Bulma
The user index now looks great again. I reworked all markup to make use
of Bulma classes and the general strucutre of other page templates.

I have a redundant page title and section title, as I want to get rid of
the page title entirely. But that will come in the next commits.

Added the option to activate a deactivated user again. I feel more
comfortable, when an Admin can do it in the software, rather than
going to the database. Maybe there is no use-case for it. But still I
don't want to lock them in, forcing them to create a new user or so.