Fixed bug in TimeStampMixin setting wrong times
Jan noticed a bug when creating users. The created at is not set to the
actual time the users were created.
Some testing revealed, that the time is set to the time the app was
started, not when the users were created.

A look in the TimeStampMixin makes this clear: We call `arrow.utcnow()`
and then get us a datetime from that. This is then our default value for
the creation and update time.
An avid python developer would spot immediately, that we only evaluate
that default value once. But I let that slip through.

Now I fixed it. I extracted the generation of the datetime into a new
function, that I then pass as the default value, without calling it. So
that it is called by SQLAlchemy everytime a default creation or update
date is needed.