Removed the "Info" column
As discussed with Pia, I removed this column. We both agree on that it
no longer holds any value.

Over time this column erroded to nothing. It used to display whether or
not a ticket was prioritized. But in a weird way, where the study
decided whether or not a ticket is prioritized. Bit convoluted. It also
opened a modal, thats doing two things at once. It should do only one
thing. So even better that it is gone.
The priorization will be added back with tags.