Removed menu items for adding users and studies
With reworking the app to tailor it to the Data Entry Staff Member I am
throwing out everything that should not be accessed by them.

Yes, this is not the way to do it. What I am trying here is to get the
app to a point, where it works as intended. Then I'll care about making
it secure, valid and the code clean.