Updated the .dockerignore
I noticed docker was sending quite a lot of data to it's build context.
I tried to remedy it with this commit, but it had no effect. However the
logs folder was the only thing not currently excluded that could have
this size.

I probably missed something. I need to look at this with more time.