Fixed logging not working
I noticed the loggin not working on the demo server. The app was
crashing as the logs folder did not exist.

This is a quick fix to just have the folder be present in the Docker
image again.

I also did the same thing to the backups folder. As I removed the
asterisk from it one or two commits ago.