Completely removed audit trail
I completely removed the old audit trail. I did that so keep the code
clean, because the audit trail was no longer properly used and needed a
complete overhaul to be fully functional and compliant again.

So in a sense I started the first step of the overhaul here.

I hope to be able to focus more on the features of this app this way and
make progress faster.