Added option to set width of page content in layout-page
Added a column that wraps the whole page content and which has a block
where one can set additional classes to it. This allows for controlling
the width of the whole pages content.

I need this as my newly designed Source-Check form doesn't look good
when it is too wide.