Completely reworked the Source-Check form
Bye table, hello card list with structured form fields.

The table was getting too small for all the fields it was holding. And
with different types of tickets on the horizont a more flecible
presenation of the form was needed.

This is just the layout however. Funktionality is not restored yet.