Rebuilt ticket list in re-check view
The old table was never really the solution we needed. With all the
infos attached to a ticket, we need a more dynamic view. This view I
tried to build here.

To be more flexible and modern in design, I opted to use a list of
cards. I tried to only display the most important information when the
card is collapsed - which is the default. The idea is that the user only
needs this information to quickly find the ticket they are looking for.

Other meta information is hidden away, and requires ucollapsing the
ticket.

Tags are also displayed in the list now! That is a big win, finally we
get them implemented.