[WIP] Started working on ticket contents in re-check list
Here I added the collapse functionality to the tickets in the re-check
list. This way, the user will be able to get all further information for
a ticket, when uncollapsing it.

For ease of programming, tickets in the list are uncollapsed by default
at the moment.

I also threw in every bit of informaion that needs to be displayed
there. Not ordered or formatted. I just threw it all in to  not forget
anything.