Evalute password conditions in place
Jan found a bug, where when pasting in a password, the hard conditions
were not displaying correctly.

After some searching, I realized, that Alpine might not like it when I
replace a whole array in my components data.

I wanted to find something similar to Vue.set. But I couldn't. So i
resorted to just updating the conditions in the array in place.