Added more advanced tag filter
The complexity here was that I now have to filter a list of
filter-values against a list of tags on a ticket.
So I refactored the filter matching algorithm to be more simplified,
instead of using a bunch of if-branching.
I now put every value, be it from the filter, or the ticket I am trying
to match, into an array if the values are not already an array. This
way I can simply see if the two arrays intersect. If they do, the filter
matches the ticket, and I don't need any kind of branching.