First basic implementation of tests using Pytest
I also opted to go with an extra `requiremetns-dev.yml` to separate dev
and prod dependencies. This way the built version of the application can
stay small.

However there is no good intgration of the prod / dev split. I just saw
this in another project and figured I want to try this as well.