More performant list filtering and sorting
I realized that I cannot access the parent components state now that I
am having two Alpine components making up the list. (ffs_list and
ffs_item.) So the code for displaying and ordering the tickets in the
list needs to stay in the markup.

However, I added some logic to make it more performant. Instead of
having each item go through the list of all items twice, in order to
determine its order and whether it should be visible or not, I added a
map that makes this easier.
I can now use the mak to quickly - and more readably - determine whether
an item should be visible. Also, I added the info about the ordinal of the
item to the item itself. This way, the ordinal can e obtained in O(1).