Tickets can be closed again
I added back this critical functionality, but i had to rework it
completely. The old way of just clicking a link was no longer viable as
it reset the list whenever the user closed a ticket.

The new way makes an async call to the backend and closes the ticket,
when successful.