Show internal ticket id in ticket comment modal
We implemented a human readable ticket id some time ago so that people
can reference tickets in some way.

This id is now shown in the header of the comment modal, replacing the
uuid that was shown there before. I'm very happy about this, as the uuid
was super long and not readable at all.