Improved Docker Compose configuration for production
In this commit, I refactored the Docker Compose configuration for
production to be less redundant.

We now have a base file, docker-compose.prod.base.yml, that holds most of the
configuration.

The other two files, docker-compose.prod.http.yml and
docker-compose.prod.https-letsencrypt.yml, are meant to extend the base
file. Thus they only hold the parts of their configuration that is
different from the base configuration.

This makes for a much less redundant Docker Compose configuration and
thus less space for making errors. (Like updating one file, but not the
other.)

I have not adapted the rest of the setup to work with this new
configuration. As I am planning to write a small script for controlling
the Docker setup for production environments, it would be in vain
anyway. The new configuration will be used by the script I am about to
write.