Optimized Dockerfile
Put the copying of the bundled frontend assets at the very end of the
Dockerfile, so that changing these assets does not trigger a reinstall
of all Python packages.

The other way around though, changing Python packages, will only trigger
the copying of the bundled frontend assets from the previous build
stage. This is a way faster operation.

All in all this should make repetetive builds faster.