First thorow of an interactive setup script
Here, I added an interactive setup and control script that should make
it easier for people to setup and manage voy.

The script was in the making for some time, but I never felt it
fulfilled some kind of committable state. So this is why it's now all
there and not built up over several commits.