Added Docker Compose as a requirement
Just to make it clear that it is required. I am under the impression,
that it now ships with Docker, but maybe I am just understanding their
CLI wrong.

Just checked it. Yeah, on Windows and MacOS it comes all together. But I
think it was always like this. On Linux it's still separate.